requirements.txt
discord.py
python-dotenv
torch
torchvision
transformers
pillow
aiohttp
requests
pytesseract
beautifulsoup4
certifi
snscrape